Update ftpHelper.java
List method added:
List Method sets the server connection to passive mode and receives the IP address and Port numbers. 
then the strings get broken into separate numbers as sub-strings delimited by ","
Last 2 Parameters of that string and port1 and port2 which are used to calculate the actual port for the passive connection.
After that the sock2 takes both the IP=hostNumberIP variable and the New Port= newport (calculated with the last 2 parameters of the returned message of the server when Using PASV) to make the new connection to be able to receive the stream of the list of directories of the Current working directory.